backend/requirements.txt
# fastapi==0.115.0
# uvicorn[standard]==0.30.0
# sqlalchemy==2.0.35
# psycopg[binary]==3.2.1
# pydantic==2.9.2
# passlib[bcrypt]==1.7.4
# PyJWT==2.9.0
# python-dotenv==1.0.1
# reportlab==4.2.2
# python-multipart==0.0.9
# langchain==0.3.0
# langchain-google-genai==2.0.0
# google-generativeai==0.8.3

# fastapi==0.115.0
# uvicorn[standard]==0.30.0
# sqlalchemy==2.0.35
# psycopg[binary]==3.2.1
# pydantic==2.9.2
# passlib[bcrypt]==1.7.4
# PyJWT==2.9.0
# python-dotenv==1.0.1
# reportlab==4.2.2
# python-multipart==0.0.9
# langchain==0.3.0
# langchain-google-genai==2.0.0
# bcrypt==4.1.2
# # google-generativeai==0.7.2
# google-generativeai==0.8.3

fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
pydantic==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
reportlab==4.2.5
python-multipart==0.0.12
bcrypt==4.2.0

# LangChain + Google Gemini (compatible combo)
langchain==0.3.4
langchain-google-genai==2.0.1
google-generativeai==0.8.3